Move flow-gen components into flow-gen folder
